requirements.txt
langchain
flask
sentence-transformers
pypdf
python-dotenv
langchain-pinecone
langchain-openai
langchain-community
langchain_huggingface
-e .